requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
cryptography==44.0.1
Django==5.1.3
django-allauth==65.4.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
drf-yasg==1.21.10
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
lucide==1.1.3
msgpack==1.1.0
optional-django==0.3.0
packaging==24.2
pillow==11.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.9.0
pyotp==2.9.0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
react==4.3.0
redis==5.2.1
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.2.3
whitenoise==6.9.0
